修改校验方式为累加和校验，将数据长度改为四个字节，TCP和解析数据调通了

+++ Task/App/include/RtuFrame.h
@@ -24,7 +24,7 @@ typedef struct
 {
     uint8_t address; //地址
     uint8_t funcode; //功能代码
-    uint8_t datalen; //数据长度
+    uint16_t datalen; //数据长度
     uint8_t* pData; //指向发送数据指针
     bool completeFlag;
 } frameRtu;
@@ -36,10 +36,10 @@ bool FrameQueneOut(uint8_t* pData);
 void  ReciveFrameDataInit(void);
 unsigned char ReciveBufferDataDealing(frameRtu* pJudgeFrame, frameRtu* pReciveFrame);
 void  GenRTUFrame(uint8_t addr, uint8_t funcode,
-                        uint8_t sendData[], uint8_t datalen, uint8_t* pRtuFrame, uint8_t *plen);
+                        uint8_t sendData[], uint32_t datalen, uint8_t* pRtuFrame, uint32_t *plen);
 void  GenRTUFrameCumulativeSum(uint8_t addr, uint8_t funcode,
              uint8_t sendData[], uint8_t datalen, uint8_t* pRtuFrame, uint8_t *plen);
-void SendFrame(uint8_t* pFrame, uint8_t len);
+void SendFrame(uint8_t* pFrame, uint32_t len);
 
 
 #ifdef	__cplusplus
